Fix code-review findings: contiguity, ImportError narrowing, bias precision, dispatch comment
- na_gemm.na_matmul(): enforce .contiguous() on both inputs so non-contiguous
  tensors (e.g. a decoded-from-LUT col-major result) don't cause the Metal
  kernel to read from wrong byte offsets; add dispatch semantics comment
  explaining threads/group_size interpretation in compile_shader.
- __init__.py: narrow bare `except ImportError: pass` to only swallow genuine
  relative-import-without-parent errors; re-raise anything else so broken
  patch dependencies surface immediately in the ComfyUI console.
- scaled_mm_fp8._mps_scaled_mm(): always add bias in f32 regardless of
  compute_dtype, restoring the old precision behaviour when compute_dtype=bf16.
- _common.decode_fp8(): skip the redundant .cpu() round-trip when the tensor
  is already on CPU and contiguous, avoiding a needless device sync.

+++ _patches/scaled_mm_fp8.py
@@ -84,7 +84,10 @@ def _mps_scaled_mm(
         out = acc.to(out.dtype)
 
     if bias is not None:
-        out = out + bias.to(out.dtype)
+        # Always add bias in f32 regardless of compute_dtype so that bf16
+        # compute does not lose precision in the bias term before the final
+        # widen (matches the old unconditional f32 bias behaviour).
+        out = out.to(torch.float32) + bias.to(torch.float32)
     if out_dtype is not None:
         out = out.to(out_dtype)
     return out
